placate  flake8
